+++ src/main/java/com/khu/gitbox/domain/member/infrastructure/MemberRepository.java
@@ -1,13 +1,15 @@
 package com.khu.gitbox.domain.member.infrastructure;
 
-import java.util.Optional;
-
+import com.khu.gitbox.domain.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
-import com.khu.gitbox.domain.member.entity.Member;
+import java.util.List;
+import java.util.Optional;
 
 @Repository
 public interface MemberRepository extends JpaRepository<Member, Long> {
-	Optional<Member> findByEmail(String email);
+    Optional<Member> findByEmail(String email);
+
+    List<Member> findByEmailIn(List<String> emails);
 }
