+++ src/main/java/com/khu/gitbox/domain/workspace/infrastructure/WorkspaceRepository.java
@@ -6,6 +6,9 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface WorkspaceRepository extends JpaRepository<Workspace, Long> {
-    Optional<Workspace> findById(Long id);
+
+    Optional<Workspace> findByOwnerId(Long ownerId);
+
+    Boolean existsByOwnerId(Long ownerId);
 }
