+++ src/main/resources/application.yaml
@@ -14,7 +14,7 @@ spring:
     database-platform: org.hibernate.dialect.MySQLDialect
     show-sql: true
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
   servlet:
     multipart:
       enabled: true
